Chore: Refactor Command Traits to Support Package
