+++ packages/admin/src/Commands/Concerns/CanIndentStrings.php
@@ -2,16 +2,13 @@
 
 namespace Filament\Commands\Concerns;
 
+use Filament\Support\Commands\Concerns\CanIndentStrings as BaseTrait;
+
+/**
+ * @deprecated Use `\Filament\Support\Commands\Concerns\CanIndentStrings` instead.
+ * @see \Filament\Support\Commands\Concerns\CanIndentStrings
+ */
 trait CanIndentStrings
 {
-    protected function indentString(string $string, int $level = 1): string
-    {
-        return implode(
-            PHP_EOL,
-            array_map(
-                fn (string $line) => str_repeat('    ', $level) . "{$line}",
-                explode(PHP_EOL, $string),
-            ),
-        );
-    }
+    use BaseTrait;
 }
