+++ packages/admin/src/Commands/MakeRelationManagerCommand.php
@@ -2,14 +2,17 @@
 
 namespace Filament\Commands;
 
+use Filament\Support\Commands\Concerns\CanIndentStrings;
+use Filament\Support\Commands\Concerns\CanManipulateFiles;
+use Filament\Support\Commands\Concerns\CanValidateInput;
 use Illuminate\Console\Command;
 use Illuminate\Support\Str;
 
 class MakeRelationManagerCommand extends Command
 {
-    use Concerns\CanIndentStrings;
-    use Concerns\CanManipulateFiles;
-    use Concerns\CanValidateInput;
+    use CanIndentStrings;
+    use CanManipulateFiles;
+    use CanValidateInput;
 
     protected $description = 'Creates a Filament relation manager class for a resource.';
