+++ packages/forms/src/Commands/MakeLayoutComponentCommand.php
@@ -2,13 +2,15 @@
 
 namespace Filament\Forms\Commands;
 
+use Filament\Support\Commands\Concerns\CanManipulateFiles;
+use Filament\Support\Commands\Concerns\CanValidateInput;
 use Illuminate\Console\Command;
 use Illuminate\Support\Str;
 
 class MakeLayoutComponentCommand extends Command
 {
-    use Concerns\CanManipulateFiles;
-    use Concerns\CanValidateInput;
+    use CanManipulateFiles;
+    use CanValidateInput;
 
     protected $description = 'Creates a form layout component class and view.';
