+++ packages/support/src/Commands/Concerns/CanIndentStrings.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Filament\Support\Commands\Concerns;
+
+trait CanIndentStrings
+{
+    protected function indentString(string $string, int $level = 1): string
+    {
+        return implode(
+            PHP_EOL,
+            array_map(
+                fn (string $line) => str_repeat('    ', $level) . "{$line}",
+                explode(PHP_EOL, $string),
+            ),
+        );
+    }
+}
